added closest distance to user map annotations and locations

+++ bikeshare-widget/BikeshareWidget/Models/Station.swift
@@ -5,6 +5,6 @@ struct Station: Identifiable {
     let id: String
     let address: String
     let bikeCapacity: Int
-    let distance: Float    
+    let distance: Double
     let coordinates: CLLocationCoordinate2D
 }
